STYLE: Merge `ImageToVectorContainerFilter` into `ImageSamplerBase`
Simplified the code by reducing the inheritance depth. (Before this commit, the chain of inheritance was: `ImageSampler` ==> `ImageSamplerBase` ==> `ImageToVectorContainerFilter` ==> `VectorContainerSource` ==> `ProcessObject`.)

Will allow using `ImageSamplerBase::m_CroppedInputImageRegion` within `ThreaderCallback`.
