+++ Common/ImageSamplers/itkImageSamplerBase.h
@@ -18,7 +18,7 @@
 #ifndef itkImageSamplerBase_h
 #define itkImageSamplerBase_h
 
-#include "itkImageToVectorContainerFilter.h"
+#include "itkVectorContainerSource.h"
 #include "itkImageSample.h"
 #include "itkVectorDataContainer.h"
 #include "itkSpatialObject.h"
@@ -40,15 +40,14 @@ namespace itk
 
 template <class TInputImage>
 class ITK_TEMPLATE_EXPORT ImageSamplerBase
-  : public ImageToVectorContainerFilter<TInputImage, VectorDataContainer<std::size_t, ImageSample<TInputImage>>>
+  : public VectorContainerSource<VectorDataContainer<std::size_t, ImageSample<TInputImage>>>
 {
 public:
   ITK_DISALLOW_COPY_AND_MOVE(ImageSamplerBase);
 
   /** Standard ITK-stuff. */
   using Self = ImageSamplerBase;
-  using Superclass =
-    ImageToVectorContainerFilter<TInputImage, VectorDataContainer<std::size_t, ImageSample<TInputImage>>>;
+  using Superclass = VectorContainerSource<VectorDataContainer<std::size_t, ImageSample<TInputImage>>>;
   using Pointer = SmartPointer<Self>;
   using ConstPointer = SmartPointer<const Self>;
 
@@ -56,17 +55,19 @@ public:
   itkNewMacro(Self);
 
   /** Run-time type information (and related methods). */
-  itkTypeMacro(ImageSamplerBase, ImageToVectorContainerFilter);
+  itkTypeMacro(ImageSamplerBase, Superclass);
 
   /** Typedefs inherited from the superclass. */
   using typename Superclass::DataObjectPointer;
   using typename Superclass::OutputVectorContainerType;
   using typename Superclass::OutputVectorContainerPointer;
-  using typename Superclass::InputImageType;
-  using typename Superclass::InputImagePointer;
-  using typename Superclass::InputImageConstPointer;
-  using typename Superclass::InputImageRegionType;
-  using typename Superclass::InputImagePixelType;
+
+  /** Some Image related typedefs. */
+  using InputImageType = TInputImage;
+  using InputImagePointer = typename InputImageType::Pointer;
+  using InputImageConstPointer = typename InputImageType::ConstPointer;
+  using InputImageRegionType = typename InputImageType::RegionType;
+  using InputImagePixelType = typename InputImageType::PixelType;
 
   /** The input image dimension. */
   itkStaticConstMacro(InputImageDimension, unsigned int, InputImageType::ImageDimension);
@@ -85,6 +86,76 @@ public:
   using MaskConstPointer = typename MaskType::ConstPointer;
   using MaskVectorType = std::vector<MaskConstPointer>;
   using InputImageRegionVectorType = std::vector<InputImageRegionType>;
+
+  /** Create a valid output. */
+  DataObject::Pointer
+  MakeOutput(unsigned int idx) override;
+
+  /** Set the input image of this process object.  */
+  void
+  SetInput(unsigned int idx, const InputImageType * input);
+
+  /** Set the input image of this process object.  */
+  void
+  SetInput(const InputImageType * input);
+
+  /** Get the input image of this process object.  */
+  const InputImageType *
+  GetInput();
+
+  /** Get the input image of this process object.  */
+  const InputImageType *
+  GetInput(unsigned int idx);
+
+  /** Get the output Mesh of this process object.  */
+  OutputVectorContainerType *
+  GetOutput();
+
+  /** Prepare the output. */
+  // virtual void GenerateOutputInformation();
+
+  /** A version of GenerateData() specific for image processing
+   * filters.  This implementation will split the processing across
+   * multiple threads. The buffer is allocated by this method. Then
+   * the BeforeThreadedGenerateData() method is called (if
+   * provided). Then, a series of threads are spawned each calling
+   * ThreadedGenerateData(). After all the threads have completed
+   * processing, the AfterThreadedGenerateData() method is called (if
+   * provided). If an image processing filter cannot be threaded, the
+   * filter should provide an implementation of GenerateData(). That
+   * implementation is responsible for allocating the output buffer.
+   * If a filter an be threaded, it should NOT provide a
+   * GenerateData() method but should provide a ThreadedGenerateData()
+   * instead.
+   *
+   * \sa ThreadedGenerateData() */
+  void
+  GenerateData() override;
+
+  /** If an imaging filter can be implemented as a multithreaded
+   * algorithm, the filter will provide an implementation of
+   * ThreadedGenerateData().  This superclass will automatically split
+   * the output image into a number of pieces, spawn multiple threads,
+   * and call ThreadedGenerateData() in each thread. Prior to spawning
+   * threads, the BeforeThreadedGenerateData() method is called. After
+   * all the threads have completed, the AfterThreadedGenerateData()
+   * method is called. If an image processing filter cannot support
+   * threading, that filter should provide an implementation of the
+   * GenerateData() method instead of providing an implementation of
+   * ThreadedGenerateData().  If a filter provides a GenerateData()
+   * method as its implementation, then the filter is responsible for
+   * allocating the output data.  If a filter provides a
+   * ThreadedGenerateData() method as its implementation, then the
+   * output memory will allocated automatically by this superclass.
+   * The ThreadedGenerateData() method should only produce the output
+   * specified by "outputThreadRegion"
+   * parameter. ThreadedGenerateData() cannot write to any other
+   * portion of the output image (as this is responsibility of a
+   * different thread).
+   *
+   * \sa GenerateData(), SplitRequestedRegion() */
+  virtual void
+  ThreadedGenerateData(const InputImageRegionType & inputRegionForThread, ThreadIdType threadId);
 
   /** ******************** Masks ******************** */
 
@@ -181,9 +252,21 @@ public:
   /** \todo: Temporary, should think about interface. */
   itkSetMacro(UseMultiThread, bool);
 
+  /** Static function used as a "callback" by the PlatformMultiThreader.  The threading
+   * library will call this routine for each thread, which will delegate the
+   * control to ThreadedGenerateData(). */
+  static ITK_THREAD_RETURN_FUNCTION_CALL_CONVENTION
+  ThreaderCallback(void * arg);
+
+  /** Internal structure used for passing image data into the threading library */
+  struct ThreadStruct //?
+  {
+    Pointer Filter;
+  };
+
 protected:
   /** The constructor. */
-  ImageSamplerBase() = default;
+  ImageSamplerBase();
 
   /** The destructor. */
   ~ImageSamplerBase() override = default;
@@ -215,11 +298,11 @@ protected:
   CropInputImageRegion();
 
   /** Multi-threaded function that does the work. */
-  void
-  BeforeThreadedGenerateData() override;
+  virtual void
+  BeforeThreadedGenerateData();
 
-  void
-  AfterThreadedGenerateData() override;
+  virtual void
+  AfterThreadedGenerateData();
 
   /***/
   unsigned long                             m_NumberOfSamples{ 0 };
@@ -229,6 +312,17 @@ protected:
   bool m_UseMultiThread{ true };
 
 private:
+  /** Split the input region into "numberOfSplits" pieces, returning
+   * region "i" as "splitRegion". This method is called "numberOfSplits" times. The
+   * regions must not overlap. The method returns the number of pieces that
+   * the routine is capable of splitting the input region,
+   * i.e. return value is less than or equal to "numberOfSplits". */
+  static unsigned int
+  SplitRegion(const InputImageRegionType & inputRegion,
+              const ThreadIdType &         threadId,
+              const ThreadIdType &         numberOfSplits,
+              InputImageRegionType &       splitRegion);
+
   /** Member variables. */
   MaskConstPointer           m_Mask{ nullptr };
   MaskVectorType             m_MaskVector{};
